pyproject.toml
[project]
name = "diting"
version = "0.13.0"
description = "DiTing - Local AI Video Knowledge Base"
readme = "README.md"
requires-python = ">=3.10"

# 核心依赖 = Web 服务运行所需的最小集
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "httpx>=0.28.0",
    "yt-dlp>=2025.12.8",
    "openai>=2.15.0",
    "dashscope>=1.14.0",
    "pydantic-settings>=2.12.0",
    "pillow>=12.0.0",
    "tomli>=2.0.1",
]

[project.optional-dependencies]
# 桌面托盘 (仅 PC 桌面需要，Docker/Web 部署不需要)
desktop = ["pystray>=0.19.5"]
# ASR Worker 基础 (torch 等重量级依赖)
worker = ["torch", "torchaudio", "torchvision", "numpy", "pyyaml"]
# 各引擎按需安装
sensevoice = ["funasr>=1.3.0", "modelscope>=1.34.0", "huggingface_hub"]
whisper = ["openai-whisper>=20250625"]
qwen = ["qwen-asr>=0.0.6"]
# 一键全装 (PC 桌面用户的默认选择)
all = ["diting[desktop,worker,sensevoice,whisper,qwen]"]
# Worker Manager GUI (独立安装向导 + 托盘管理)
manager = ["customtkinter>=5.2.0", "darkdetect>=0.8.0", "pystray>=0.19.5", "pyinstaller>=6.19.0"]
# Tauri 桌面端打包
tauri = ["pyinstaller>=6.19.0"]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
explicit = true  # 这一行非常重要，它防止非指定的包去这里找，也防止这里的包去别处找

[tool.uv.sources]
torch = { index = "pytorch-gpu" }
torchvision = { index = "pytorch-gpu" }
torchaudio = { index = "pytorch-gpu" }
